#44 - expanding into the auto salesprices

+++ OneSila/currencies/helpers.py
@@ -1,4 +1,5 @@
 import math
+from decimal import Decimal
 
 
 def currency_convert(round_prices_up_to, exchange_rate, price):
@@ -11,7 +12,7 @@ def currency_convert(round_prices_up_to, exchange_rate, price):
     if price is None:
         new_price = None
     else:
-        new_price = (price * exchange_rate)
+        new_price = (price * Decimal(exchange_rate))
 
     new_price = roundup(new_price, round_prices_up_to)
 
@@ -21,4 +22,4 @@ def currency_convert(round_prices_up_to, exchange_rate, price):
 def roundup(x, ceil):
     ''' Roundup a number to the nearest ceiling
     For example:  roundup(111, 30) returns 120'''
-    return int(math.ceil(x / float(ceil))) * ceil
+    return int(math.ceil(x / Decimal(ceil))) * Decimal(ceil)
